feat: Moved matched filter to mtmf.py file
